% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country details}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 174 rows and 11 columns.}
\source{
Comtrade \url{http://comtrade.un.org}
}
\usage{
countries
}
\description{
A data frame of 174 records and 11 variables containing country metadata.
}
\details{
\describe{
\item{stuff to do}{...}

}
}
\keyword{datasets}
