% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Country details}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 246 rows and 4 columns.}
\source{
Comtrade \url{http://comtrade.un.org}
}
\usage{
country_codes
}
\description{
A data frame country names, ISO codes and UN codes.
}
\details{
\describe{
\item{stuff to do}{...}

}
}
\keyword{datasets}
