% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ffd_indicators}
\alias{ffd_indicators}
\title{Trade indicators}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1378 rows and 14 columns.}
\source{
Comtrade \url{http://comtrade.un.org}
}
\usage{
ffd_indicators
}
\description{
A data frame country trade indicators.
}
\details{
\describe{
\item{stuff to do}{...}

}
}
\keyword{datasets}
