% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_country_flag.R
\name{mod_country_flag_ui}
\alias{mod_country_flag_ui}
\alias{mod_country_flag_server}
\title{mod_country_flag_ui and mod_country_flag_server}
\usage{
mod_country_flag_ui(id)

mod_country_flag_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
