% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_country_map.R
\name{mod_country_map_ui}
\alias{mod_country_map_ui}
\alias{mod_country_map_server}
\title{mod_country_map_ui and mod_country_map_server}
\usage{
mod_country_map_ui(id)

mod_country_map_server(input, output, session, country)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{country}{a reactive containing a country ISO3 code}

\item{height}{in pixels or percentage}

\item{width}{in pixels or percentage}
}
\description{
A shiny Module.
}
\keyword{internal}
