% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_country_name.R
\name{mod_country_name_ui}
\alias{mod_country_name_ui}
\alias{mod_country_name_server}
\title{mod_country_name_ui and mod_country_name_server}
\usage{
mod_country_name_ui(id)

mod_country_name_server(input, output, session, country)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
