% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ffd_country_series.R
\name{mod_ffd_country_series_ui}
\alias{mod_ffd_country_series_ui}
\alias{mod_ffd_country_series_server}
\title{mod_ffd_country_series_ui and mod_ffd_country_series_server}
\usage{
mod_ffd_country_series_ui(id)

mod_ffd_country_series_server(input, output, session, country)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
