% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ffd_indicator_table.R
\name{mod_ffd_indicator_table_ui}
\alias{mod_ffd_indicator_table_ui}
\alias{mod_ffd_indicator_table_server}
\title{mod_ffd_indicator_table_ui and mod_ffd_indicator_table_server}
\usage{
mod_ffd_indicator_table_ui(id)

mod_ffd_indicator_table_server(input, output, session, country)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{a}{reactive containing a country ISO3 code}
}
\value{
a rendered reactive DT datatable
}
\description{
Shiny module. Produces a DT datatable of trade indicators for
  the Market overview section.
}
\keyword{internal}
