% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_test.R
\name{mod_test_ui}
\alias{mod_test_ui}
\alias{mod_test_server}
\title{mod_test_ui and mod_test_server}
\usage{
mod_test_ui(id)

mod_test_server(input, output, session, country)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
