% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wb_indicator_text.R
\name{mod_wb_indicator_text_ui}
\alias{mod_wb_indicator_text_ui}
\alias{mod_wb_indicator_text_server}
\title{mod_wb_indicator_text_ui and mod_wb_indicator_text_server}
\usage{
mod_wb_indicator_text_ui(id)

mod_wb_indicator_text_server(input, output, session, country, indicator)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
