% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_indicators}
\alias{wb_indicators}
\title{World Bank indicators}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 184 rows and 10 columns.}
\source{
Comtrade \url{http://comtrade.un.org}
}
\usage{
wb_indicators
}
\description{
A data frame World Bank indicators.
}
\details{
\describe{
\item{stuff to do}{...}

}
}
\keyword{datasets}
