% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_country_flag.R
\name{mod_country_flag_ui}
\alias{mod_country_flag_ui}
\alias{mod_country_flag_server}
\title{mod_country_flag_ui and mod_country_flag_server}
\usage{
mod_country_flag_ui(id)

mod_country_flag_server(input, output, session, dataset, country,
  height = "100\%", width = "100\%")
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataset, }{dataset (non reactive) containing at least reporter_iso and
urls to country flag images. Currently expect this dataset to have
'reporter_iso' and 'png'. Should change this to specify using tidy
evaluation in the future.}

\item{country, }{list containing reactive country name (reporter_iso) to
filter on}

\item{height, }{text containing either percentage ("50%") or pixel size
("640")}

\item{width, }{text containing either percentage ("50%") or pixel size ("640")}
}
\description{
A shiny Module.
}
\keyword{internal}
